Added style to the fourth element in index.html.
Added id "fourth" to the fourth element in the
index.html and in the style.css, added light green
background, font size of 24px and bold font weight
to that element using the id selector. Also, as
the fourth element shared the same font size as
the third odd element, grouped them together for
the font size.