Fixed the style of the smaller paragraph in style.css.
In index.html, the paragraph supposed to be small
in appearance had both the classes "para" and
"small-para". In style.css, the style declaration
for "para" class was after that of "small-para"
class. So, the style declaration for para class
with a larger font size took the precedence,
preventing smaller font size to be applied to the
paragraph with both the classes, as both the
declarations had the same specificity of just one
class. The style declaration for the "para" class
was brought before that of "small-para" class. So
that, for the paragraph with "small-para" class,
smaller font size be applied than the paragraphs
with just "para" class.