Fixed style of the confirm button.
In the index.html, the confirm button had the
classes "button" and "confirm". In style.css, the
style declaration for the "button" class was after
the "confirm" class, so the style declaration for
the "button" class took the precedence and general
button style was applied to the confirm button,
preventing it from getting a green background,
white text, and bold text weight. So, the style
declaration for the "confirm" class was placed
after the "button" class, so that the confirm
button will have its unique style over the general
style of the "button" class.