Fixed style for the smaller child div.
In index.html, a div has classes "text" and
"child", which is to be smaller than other texts
in size and with extra bold text. In style.css,
there are style declaration for class "child" and
the combination of the "div" type selector and
".text" class selector in this order. Because of
which the specificity of the combination is higher
than of ".child" class selector alone, so the
general style for the div text is applied applied
to the child div as well which is supposed to be
smaller and in bold. So, the ".child" selector was
combined with "div" type selector and ".text"
class selector, so that it would regain its
desired style.