Added style to the ".title" class selector.
In the index.html, the first child of the card div
element is the div element with ".title" class.
In the style.css, added font size of 1 rem and
padding of 8px to the ".title" class selector.
The padding of 8px added 8px of space around the
text content of the title div element inside its
border, and the font size of 1rem set the font
size of its text content to 16px.