Added style to the ".content" class selector.
Added bottom margin of 8px and padding of 12px
to the ".content" class selector. In the
index.html, the second child of the card div
element has the class ".content". In style.css,
added the bottom margin, so that there is 8px
of space below that div element outside its border
and the padding so that there is 12px of space
around the text content of the div element inside
its border.