Added style to the button container and the button.
Added padding block of 8px to the button container
and aligned its contents centrally. Added margin
inline auto, inline padding of 20px and block
padding of 8px to the button element and set its
display to "block". In index.html, the third child
of card div element has class "button-container"
which contains text with a button element. In
style.css, inside the ".button-container" class
selector, the padding block added 8px of space
above and below its content inside its border, and
the "text-align" to center aligned its content
centrally. And in the ".button" type selector,
changing its display to "block" set it as a
block-level element from its default inline. The
margin-inline auto distributed any available space
equally between the start and end sides of the
button. The inline padding added space of20px at
the left and right sides inside the border of the
button, and block padding added space of 8px at
the top and bottom sides inside the border of the
button.