Added flex properties to the icon div.
Added flex-growth of 1, flex-shrink of 0 and
flex-basis "auto" inside the .icon class selector
for the icon div. The flex-growth of 1 for the
icon ensured that it grows in size evenly relative
to the other flex items. The flex-shrink of 0
ensured that the icon will not shrink in size
when the window is minimized. The flex-basis of
auto ensured that the icon will use its defined
width every time.