Added style to the header div.
The header div contains confirmation text and a
close button. Set the header div as a flex
container by adding display "flex" to the .header
class selector. Then using justify-content
"space-between" aligned the flex items of the
header div to have space in between each other
along the horizontal axis and using align-items
"center" aligned the flex items centrally across
the vertical axis. Increased the font-size of the
header text to 18px and set the font-weight to
bold.