Aligned the body contents centrally, removed bullets from the links, and added style to the header div for the header links.
Using text-align "center" to the body type
selector, aligned all the body elements in the
middle portion of the page. Set list-style-type
to "none" in the ul type selector, which removed
default bullets in the front of the list items
(links). In the ".header" class selector, added
display "flex" to set the header as a flex
container, and used justify-content
"space-between" to add space between the right and
the left links, and used align-items "center" to
position right and left links centrally across the
vertical axis inside the header div.