Removed the text decoration from and changed the color of the link and added style to the header links.
In the "a" type selector, set text-decoration to
"none" which removed the default underline from
the link and set its color to #646060 which sets
the links' color to light gray. Using descendant
combinator ".header" .left-links" and ".header
. right-links", set the right-links and left-links
div as flex container, aligned it centrally inside
those divs across the horizontal axis, added gap
of 18px to have that space between left and right
links, and added padding-inline of 16px, so that
the links aren't touching the edge of the window.