Set body and content as a flex container with flex-direction of column, and added padding to the buttons.
Adding display "flex" and flex-direction "column"
inside the "body" type selector and ".content"
class selector. In the ".content" class selector,
aligned its flex items (logo, text input, and
buttons) centrally across the vertical and
horizontal axis inside the content div using
justify-content and align-items "center". Added
padding-block of 6px to the buttons which added
6px of space to their top and bottom, and padding-
inline of 12px which added 12px of space to the
right and left of the buttons.