Set the body as the flex container.
Added display "flex" inside the body type selector
and set the flex-direction to "column". The
display property set the body as flex containers
and the elements inside the body as flex items,
and the flex direction of column stacked the
flex-items on top of each other, like in a
column.