Set the content div as a flex container and styled the side bar links.
Inside the ".content" class selector, added
display "flex". In the "ul" type selector, changed
the list style type to none which removed the
bullet points, and changed the default
padding-inline-start and margin-block to 0px and
added margin-block-start of 8px. In the "li" type
selector, added padding of 6px. In the "a" type
selector, set text-decoration to none which
removed the default underline underneath the links
and changed the links color to #fff (white).