Set flex-wrap property for the container div and added padding of 48px around the cards container. Set flex-shrink property for the sidebar and added width, height, margin and padding to the cards.
Added flex-wrap "wrap" inside the ".container"
class selector. These let the card take up
multiple lines as the window expands by filling
up the available space, and stack when the window
is minimized. Added padding of 48px to add that
space around the cards container. Set flex-shrink
to 0 for the side bar so that the side bar will
maintain its original width even when the window
shrinks in size. Added width of 250px and height
of 300px to the cards and margin of 15px and
padding of 13px.