Added flex property to header, content, and footer and typographic changes to the header and footer.
Added flex property of 1 1 auto to the header,
content, and the footer. The flex basis auto
use the defined height and width of those divs.
Set header and footer as flex container and
header's text was vertically centered and footer's
text vertically and horizontally centered. Added
font size of 32px, and font-weight of 900 to the
header.